[IMP] payment_stripe: move stripe_options.js to interactions
X-original-commit: b196582a8eed17bbdb4db875edb9b6e7a40e1dd0
