+++ addons/payment_stripe/static/src/interactions/stripe_options.js
